python/requirements.txt
watchdog>=6.0.0
pdfplumber>=0.11.0
python-docx>=1.1.0
openpyxl>=3.1.0
qdrant-client>=1.17.0
ollama>=0.4.0
langchain-text-splitters>=0.3.0
